Separate MegaRAID-related logic to megaraid.py
