+++ smartprom.py
@@ -8,6 +8,8 @@ from typing import Tuple
 
 import prometheus_client
 
+import megaraid
+
 LABELS = ['drive', 'type', 'model_family', 'model_name', 'serial_number']
 DRIVES = {}
 METRICS = {}
@@ -16,7 +18,6 @@ METRICS = {}
 SAT_TYPES = ['sat', 'usbjmicron', 'usbprolific', 'usbsunplus']
 NVME_TYPES = ['nvme', 'sntasmedia', 'sntjmicron', 'sntrealtek']
 SCSI_TYPES = ['scsi']
-MEGARAID_TYPE_PATTERN = r"(sat\+)?(megaraid,\d+)"
 
 
 def run_smartctl_cmd(args: list) -> Tuple[str, int]:
@@ -60,14 +61,13 @@ def get_drives() -> dict:
         devices = result_json['devices']
         for device in devices:
             dev = device["name"]
-            if re.match(MEGARAID_TYPE_PATTERN, device["type"]):
+            if re.match(megaraid.MEGARAID_TYPE_PATTERN, device["type"]):
                 # If drive is connected by MegaRAID, dev has a bus name like "/dev/bus/0".
                 # After retrieving the disk information using the bus name,
                 # replace dev with a disk ID such as "megaraid,0".
-                bus_name = dev
-                disk_attrs = get_megaraid_device_info(device["type"], bus_name)
-                disk_attrs["bus_device"] = bus_name
-                disk_attrs["megaraid_id"] = get_megaraid_device_id(device["type"])
+                disk_attrs = megaraid.get_megaraid_device_info(dev, device["type"])
+                disk_attrs["bus_device"] = dev
+                disk_attrs["megaraid_id"] = megaraid.get_megaraid_device_id(device["type"])
                 dev = disk_attrs["megaraid_id"]
 
                 # Generate device["type"] from device["protocol"]
@@ -94,48 +94,6 @@ def get_device_info(dev: str) -> dict:
         'model_name': results.get("model_name", "Unknown"),
         'serial_number': results.get("serial_number", "Unknown")
     }
-
-
-def get_megaraid_device_info(typ: str, dev: str) -> dict:
-    """
-    Get device information connected with MegaRAID,
-    and process the information into get_device_info compatible format.
-    """
-    megaraid_id = get_megaraid_device_id(typ)
-    if megaraid_id is None:
-        return {}
-
-    results, _ = run_smartctl_cmd(
-        ["smartctl", "-i", "--json=c", "-d", megaraid_id, dev]
-    )
-    results = json.loads(results)
-    serial_number = results.get("serial_number", "Unknown")
-    model_family = results.get("model_family", "Unknown")
-
-    # When using SAS drive and smartmontools r5286 and later,
-    # scsi_ prefix is added to model_name field.
-    # https://sourceforge.net/p/smartmontools/code/5286/
-    model_name = results.get(
-        "scsi_model_name",
-        results.get("model_name", "Unknown"),
-    )
-
-    return {
-        "model_family": model_family,
-        "model_name": model_name,
-        "serial_number": serial_number,
-    }
-
-
-def get_megaraid_device_id(typ: str) -> str | None:
-    """
-    Returns the device ID on the MegaRAID from the typ string
-    """
-    megaraid_match = re.search(MEGARAID_TYPE_PATTERN, typ)
-    if not megaraid_match:
-        return None
-
-    return megaraid_match.group(2)
 
 
 def get_smart_status(results: dict) -> int:
@@ -245,32 +203,6 @@ def results_to_attributes_scsi(data: dict) -> dict:
     return attributes
 
 
-def smart_megaraid(megaraid_id: str, dev: str) -> dict:
-    """
-    Runs the smartctl command on device connected by MegaRAID
-    and processes its attributes
-    """
-    results, exit_code = run_smartctl_cmd(
-        ["smartctl", "-A", "-H", "-d", megaraid_id, "--json=c", dev]
-    )
-    results = json.loads(results)
-
-    if results["device"]["protocol"] == "ATA":
-        # SATA device on MegaRAID
-        data = results["ata_smart_attributes"]["table"]
-        attributes = table_to_attributes_sat(data)
-        attributes["smart_passed"] = (0, get_smart_status(results))
-        attributes["exit_code"] = (0, exit_code)
-        return attributes
-    elif results["device"]["protocol"] == "SCSI":
-        # SAS device on MegaRAID
-        attributes = results_to_attributes_scsi(results)
-        attributes["smart_passed"] = get_smart_status(results)
-        attributes["exit_code"] = exit_code
-        return attributes
-    return {}
-
-
 def collect():
     """
     Collect all drive metrics and save them as Gauge type
@@ -281,8 +213,8 @@ def collect():
         typ = drive_attrs['type']
         try:
             if "megaraid_id" in drive_attrs:
-                attrs = smart_megaraid(
-                    drive_attrs["megaraid_id"], drive_attrs["bus_device"]
+                attrs = megaraid.smart_megaraid(
+                    drive_attrs["bus_device"], drive_attrs["megaraid_id"]
                 )
             elif typ in SAT_TYPES:
                 attrs = smart_sat(drive)
